feat(build): remove all unused import & unused variable
