+++ frontend/src/components/ui/sidebar/expendable.tsx
@@ -1,4 +1,4 @@
-import { faArrowDown, faChevronDown, faHome } from "@fortawesome/free-solid-svg-icons";
+import { faChevronDown, faHome } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 
 export default function Expendable() {
